requirements.txt
alembic==0.9.7
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cryptography==2.1.4
Flask==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
gevent==1.2.2
greenlet==0.4.13
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
oauthlib==2.0.6
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
SQLAlchemy==1.2.2
urllib3==1.22
Werkzeug==0.14.1
